refactor(web): render the notifications card from one component
The Home tab and the Notifications tab each carried their own ~50-line
copy of the push-permission ladder, differing only in how they treat the
desktop app. Extracted into settings/SettingsNotifications.vue, which
owns the push state it needs; the Home tab passes hide-in-desktop-app to
keep leaving that surface to the tray.

Card scaffolding CSS is mirrored locally, as SettingsAutomation.vue
already does — SettingsView's styles are scoped and do not reach into a
child's inner elements.
